+++ projects/np-ui-lib/src/lib/np-auto-complete/np-auto-complete.module.ts
@@ -3,7 +3,7 @@ import { CommonModule } from "@angular/common";
 import { OverlayModule } from "@angular/cdk/overlay";
 import { PortalModule } from "@angular/cdk/portal";
 import { NpAutoCompleteComponent } from "./np-auto-complete.component";
-import { NpTranslationsModule } from "../np-translations/np-tranlations.module";
+import { NpTranslationsModule } from "../np-translations/np-translations.module";
 import { NpAutofocusModule } from "../np-utility/np-autofocus.module";
 import { NpOrderByModule } from "../np-utility/np-orderby.module";
 import { NpHightlightModule } from "../np-utility/np-highlight.module";