+++ projects/np-ui-lib/src/lib/np-date-picker/np-date-picker.module.ts
@@ -4,7 +4,7 @@ import { NpDatePickerComponent } from "./np-date-picker.component";
 import { NpTooltipModule } from "../np-tooltip/np-tooltip.module";
 import { OverlayModule } from "@angular/cdk/overlay";
 import { PortalModule } from "@angular/cdk/portal";
-import { NpTranslationsModule } from "../np-translations/np-tranlations.module";
+import { NpTranslationsModule } from "../np-translations/np-translations.module";
 import { NpAutofocusModule } from "../np-utility/np-autofocus.module";
 
 @NgModule({