+++ projects/np-ui-lib/src/lib/np-paginator/np-paginator.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from "@angular/core";
 import { CommonModule } from "@angular/common";
 import { NpPaginatorComponent } from "./np-paginator.component";
-import { NpTranslationsModule } from "../np-translations/np-tranlations.module";
+import { NpTranslationsModule } from "../np-translations/np-translations.module";
 
 @NgModule({
   declarations: [NpPaginatorComponent],