+++ projects/np-ui-lib/src/lib/np-tags/np-tags.module.ts
@@ -3,7 +3,7 @@ import { CommonModule } from "@angular/common";
 import { NpTagsComponent } from "./np-tags.component";
 import { OverlayModule } from "@angular/cdk/overlay";
 import { PortalModule } from "@angular/cdk/portal";
-import { NpTranslationsModule } from "../np-translations/np-tranlations.module";
+import { NpTranslationsModule } from "../np-translations/np-translations.module";
 import { NpAutofocusModule } from "../np-utility/np-autofocus.module";
 import { NpOrderByModule } from "../np-utility/np-orderby.module";
 import { NpHightlightModule } from "../np-utility/np-highlight.module";