+++ projects/np-ui-lib/src/lib/np-time-picker/np-time-picker.module.ts
@@ -3,7 +3,7 @@ import { NpTimePickerComponent } from "./np-time-picker.component";
 import { CommonModule } from "@angular/common";
 import { OverlayModule } from "@angular/cdk/overlay";
 import { PortalModule } from "@angular/cdk/portal";
-import { NpTranslationsModule } from "../np-translations/np-tranlations.module";
+import { NpTranslationsModule } from "../np-translations/np-translations.module";
 import { NpAutofocusModule } from "../np-utility/np-autofocus.module";
 
 @NgModule({