+++ projects/np-ui-lib/src/public-api.ts
@@ -192,7 +192,7 @@ export * from "./lib/np-textarea/np-textarea.module";
 export * from "./lib/np-textarea/np-textarea.component";
 
 /*Translations */
-export * from "./lib/np-translations/np-tranlations.module";
+export * from "./lib/np-translations/np-translations.module";
 export * from "./lib/np-translations/np-translate.pipe";
 export * from "./lib/np-translations/np-translations.service";